perf(k3): the bf16 elementwise family streams eight columns a thread
add2, mul_sigmoid, situ and o_norm_gate were TileLang batched kernels
walking one element per thread with 2-byte loads and stores — 1–2 TB/s on
the chunked-prefill rows, 220 ms of the CP8 128k deep rank (add2 91,
o_norm_gate 75, situ 36, mul_sigmoid 21) for what is 6 TB/s of traffic.

They are hand-written CUDA now, sixteen bytes per thread per operand and
shape-agnostic, with the retired kernels' arithmetic kept term for term:
TileLang's bfloat16_t is cutlass's, whose `+` is __hadd, `*` is __hmul
and cast is cvt.rn, so the bf16x2 forms round identically; the f32 chains
keep expf/tanhf/rsqrtf and IEEE division under the same -O3 without
fast-math. o_norm_gate's 128-wide xor butterfly (64 and 32 through shared
memory, 16..1 by shuffle) is reproduced pair for pair by sixteen lanes of
eight columns — lane xor 8/4/2/1 then slots j^4/j^2/j^1 — so the norm
scale is the same bits in every lane. eps is a runtime argument
(K3_RMS_EPS at the one call site) instead of a compiled constant.

Forty TileLang instantiations retire with the family (280 -> 240).

Gates (pruned 224-expert checkpoint): golden_decode 13/13 exact — decode
runs the same add2/mul_sigmoid/situ — paged_kv 3/3, spec_verify 6/6,
cp_prefill matches + sweep, CP4 vs CP1 at 64k rel_l2 unchanged to four
digits (8.706e-2 / 6.991e-2). CP4 64k in-process wall 3,785 -> 3,639 ms.
