+++ pegainfer-kernels/src/ffi/k3.rs
@@ -189,6 +189,54 @@ unsafe extern "C" {
         stream: CUstream,
     ) -> CUresult;
 
+    /// bf16 elementwise family (`csrc/k3/k3_elementwise.cu`), eight columns
+    /// per thread, bit-identical to the retired TileLang batched kernels.
+    /// `O = A + Bt` in bf16 addition, all `[b, n]`; n a multiple of 8.
+    pub fn k3_add2_cuda(
+        a: *const c_void,
+        bt: *const c_void,
+        o: *mut c_void,
+        b: i32,
+        n: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    /// `O = A * bf16(sigmoid(Bt))`, the MLA sigmoid output gate. All `[b, n]`.
+    pub fn k3_mul_sigmoid_cuda(
+        a: *const c_void,
+        bt: *const c_void,
+        o: *mut c_void,
+        b: i32,
+        n: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    /// K3's situ activation `4*tanh(g/4)*sigmoid(g) * 25*tanh(u/25)`, computed
+    /// in f32 and landed bf16 once. All `[b, n]`.
+    pub fn k3_situ_cuda(
+        g: *const c_void,
+        u: *const c_void,
+        o: *mut c_void,
+        b: i32,
+        n: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    /// `kda_core`'s tail on its own: per (row, head) f32 rms_norm of the bf16
+    /// attention landing `X` times the o_norm gamma `Go [128]`, landed once,
+    /// times the bf16 sigmoid of the output gate `G2`. head_dim must be 128.
+    pub fn k3_o_norm_gate_cuda(
+        x: *const c_void,
+        g2: *const c_void,
+        go: *const f32,
+        out: *mut c_void,
+        b: i32,
+        num_heads: i32,
+        head_dim: i32,
+        eps: f32,
+        stream: CUstream,
+    ) -> CUresult;
+
     /// Chunked-prefill conv + silu over one q/k/v stream of one segment
     /// (`csrc/k3/k3_conv_silu_chunk.cu`): `p [tokens, inner]` f32 partial,
     /// `cw [4, inner]` f32 taps, `carry [3, inner]` bf16 window preceding
