+++ pegainfer-kernels/src/ffi/k3_tilelang.rs
@@ -47,37 +47,6 @@ unsafe extern "C" {
         stream: CUstream,
     ) -> i32;
 
-    /// `O = A + Bt` in bf16 addition, all `[b, n]`.
-    pub fn k3_add2_batched(
-        a: *const c_void,
-        bt: *const c_void,
-        o: *mut c_void,
-        b: i32,
-        n: i32,
-        stream: CUstream,
-    ) -> i32;
-
-    /// `O = A * bf16(sigmoid(Bt))`, the MLA sigmoid output gate. All `[b, n]`.
-    pub fn k3_mul_sigmoid_batched(
-        a: *const c_void,
-        bt: *const c_void,
-        o: *mut c_void,
-        b: i32,
-        n: i32,
-        stream: CUstream,
-    ) -> i32;
-
-    /// K3's situ activation `4*tanh(g/4)*sigmoid(g) * 25*tanh(u/25)`, computed
-    /// in f32 and landed bf16 once. All `[b, n]`.
-    pub fn k3_situ_batched(
-        g: *const c_void,
-        u: *const c_void,
-        o: *mut c_void,
-        b: i32,
-        n: i32,
-        stream: CUstream,
-    ) -> i32;
-
     /// Causal depthwise convolution over the `width`-slot window plus silu.
     /// `P [b, split_k, kp]` f32 partials land into `X [b, kp]` bf16, which is
     /// also the newest window slot; `Cs`/`Sn [b, width - 1, kp]` are the carried
@@ -120,20 +89,6 @@ unsafe extern "C" {
         num_heads: i32,
         head_dim: i32,
         split_k_gate: i32,
-        stream: CUstream,
-    ) -> i32;
-
-    /// `kda_core`'s tail on its own: per (row, head) f32 rms_norm of the bf16
-    /// attention landing `X` times the o_norm gamma `Go [head_dim]`, landed
-    /// once, times the bf16 sigmoid of the output gate `G2`. eps compiled in.
-    pub fn k3_o_norm_gate_batched(
-        x: *const c_void,
-        g2: *const c_void,
-        go: *const f32,
-        out: *mut c_void,
-        b: i32,
-        num_heads: i32,
-        head_dim: i32,
         stream: CUstream,
     ) -> i32;
