+++ pegainfer-kernels/src/ops/k3_tilelang.rs
@@ -211,119 +211,6 @@ pub fn k3_land_rms_norm_rbs_batched_launch(
     )
 }
 
-/// The two `(A, Bt) -> O` elementwise kernels share a signature; only the
-/// launcher symbol and the error label differ.
-type BinaryLauncher = unsafe extern "C" fn(
-    *const c_void,
-    *const c_void,
-    *mut c_void,
-    i32,
-    i32,
-    cudarc::driver::sys::CUstream,
-) -> i32;
-
-fn k3_binary_batched_launch(
-    ctx: &DeviceContext,
-    launcher: BinaryLauncher,
-    what: &str,
-    b: usize,
-    n: usize,
-    a: &CudaSlice<bf16>,
-    bt: &CudaSlice<bf16>,
-    o: &mut CudaSlice<bf16>,
-) -> Result<()> {
-    check_bucket(b)?;
-    ensure!(
-        a.len() >= b * n && bt.len() >= b * n && o.len() >= b * n,
-        "K3 {what} buffers too small for b={b}, n={n}: a {}, bt {}, o {}",
-        a.len(),
-        bt.len(),
-        o.len()
-    );
-    let (a_ptr, _a_guard) = a.device_ptr(&ctx.stream);
-    let (bt_ptr, _bt_guard) = bt.device_ptr(&ctx.stream);
-    let (o_ptr, _o_guard) = o.device_ptr_mut(&ctx.stream);
-    let rc = unsafe {
-        launcher(
-            a_ptr as *const c_void,
-            bt_ptr as *const c_void,
-            o_ptr as *mut c_void,
-            b as i32,
-            n as i32,
-            ctx.stream.cu_stream(),
-        )
-    };
-    check(rc, &format!("K3 {what} (B={b}, N={n})"))
-}
-
-/// `o = a + bt`, added in bf16 — the residual adds, and routed + shared.
-pub fn k3_add2_batched_launch(
-    ctx: &DeviceContext,
-    b: usize,
-    n: usize,
-    a: &CudaSlice<bf16>,
-    bt: &CudaSlice<bf16>,
-    o: &mut CudaSlice<bf16>,
-) -> Result<()> {
-    k3_binary_batched_launch(ctx, ffi::k3_add2_batched, "add2_batched", b, n, a, bt, o)
-}
-
-/// `o = a * bf16(sigmoid(bt))`, the MLA sigmoid output gate. The sigmoid is
-/// taken in f32 and lands in bf16 before the product.
-pub fn k3_mul_sigmoid_batched_launch(
-    ctx: &DeviceContext,
-    b: usize,
-    n: usize,
-    a: &CudaSlice<bf16>,
-    bt: &CudaSlice<bf16>,
-    o: &mut CudaSlice<bf16>,
-) -> Result<()> {
-    k3_binary_batched_launch(
-        ctx,
-        ffi::k3_mul_sigmoid_batched,
-        "mul_sigmoid_batched",
-        b,
-        n,
-        a,
-        bt,
-        o,
-    )
-}
-
-/// K3's situ activation: `4*tanh(g/4)*sigmoid(g) * 25*tanh(u/25)` in f32,
-/// landing bf16 once. The two betas are compiled in.
-pub fn k3_situ_batched_launch(
-    ctx: &DeviceContext,
-    b: usize,
-    n: usize,
-    g: &CudaSlice<bf16>,
-    u: &CudaSlice<bf16>,
-    o: &mut CudaSlice<bf16>,
-) -> Result<()> {
-    check_bucket(b)?;
-    ensure!(
-        g.len() >= b * n && u.len() >= b * n && o.len() >= b * n,
-        "K3 situ buffers too small for b={b}, n={n}: g {}, u {}, o {}",
-        g.len(),
-        u.len(),
-        o.len()
-    );
-    let (g_ptr, _g_guard) = g.device_ptr(&ctx.stream);
-    let (u_ptr, _u_guard) = u.device_ptr(&ctx.stream);
-    let (o_ptr, _o_guard) = o.device_ptr_mut(&ctx.stream);
-    let rc = unsafe {
-        ffi::k3_situ_batched(
-            g_ptr as *const c_void,
-            u_ptr as *const c_void,
-            o_ptr as *mut c_void,
-            b as i32,
-            n as i32,
-            ctx.stream.cu_stream(),
-        )
-    };
-    check(rc, &format!("K3 situ_batched (B={b}, N={n})"))
-}
-
 /// KDA short convolution plus silu, one token per row.
 ///
 /// `p` is the projection's f32 partial; its bf16 landing is written to `x` and
@@ -491,53 +378,6 @@ pub fn k3_kda_core_batched_launch(
     check(
         rc,
         &format!("K3 kda_core_batched (B={b}, KH={num_heads}, KD={head_dim})"),
-    )
-}
-
-/// `kda_core`'s tail on its own: per (row, head) the f32 rms_norm of the bf16
-/// attention landing `x` times the o_norm gamma `go [head_dim]`, landed once,
-/// times the bf16 sigmoid of the output gate `g2` — word-for-word the batched
-/// kernel's last loop. Chunked prefill computes the attention through FlashKDA
-/// and finishes its rows here.
-pub fn k3_o_norm_gate_batched_launch(
-    ctx: &DeviceContext,
-    b: usize,
-    num_heads: usize,
-    head_dim: usize,
-    x: &CudaSlice<bf16>,
-    g2: &CudaSlice<bf16>,
-    go: &CudaSlice<f32>,
-    out: &mut CudaSlice<bf16>,
-) -> Result<()> {
-    check_bucket(b)?;
-    let kp = num_heads * head_dim;
-    ensure!(
-        x.len() >= b * kp && g2.len() >= b * kp && out.len() >= b * kp && go.len() >= head_dim,
-        "K3 o_norm_gate buffers too small for b={b}, kp={kp}: x {}, g2 {}, go {}, out {}",
-        x.len(),
-        g2.len(),
-        go.len(),
-        out.len()
-    );
-    let (x_ptr, _x_guard) = x.device_ptr(&ctx.stream);
-    let (g2_ptr, _g2_guard) = g2.device_ptr(&ctx.stream);
-    let (go_ptr, _go_guard) = go.device_ptr(&ctx.stream);
-    let (out_ptr, _out_guard) = out.device_ptr_mut(&ctx.stream);
-    let rc = unsafe {
-        ffi::k3_o_norm_gate_batched(
-            x_ptr as *const c_void,
-            g2_ptr as *const c_void,
-            go_ptr as *const f32,
-            out_ptr as *mut c_void,
-            b as i32,
-            num_heads as i32,
-            head_dim as i32,
-            ctx.stream.cu_stream(),
-        )
-    };
-    check(
-        rc,
-        &format!("K3 o_norm_gate_batched (B={b}, KH={num_heads}, KD={head_dim})"),
     )
 }
